net: zstream: Add zstream_writeall() convenience function
This function is useful for blocking streams, to account for
possibility of short writes. This function will write exactly
as many bytes as requested by user, by repeating the operation
with remaining data in case as short write(s) happen, unless
an error happens before all data is written. The number of bytes
actually written is returned using the out-pointer argument
(it can be smaller than requested if error occurred).

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>