samples: net: google_iot_device: bsp: iotc_bsp_io_net_zephyr: Use bare constants POLLIN/POLLOUT/POLLERR.
The rest of this file uses bare POSIX names, not "zsock_" prefixed
variants, so it only makes sense to do the same for these constants.

And next step of porting the sample to native Zephyr targets is to
enable POSIX subsystem (CONFIG_POSIX_API), which only exports standard
POSIX name, so it's required for it too.

This changes should be squashed into the original commit introducing iotc_bsp_io_net_zephyr.c.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>